requirements.txt
openai
faiss-cpu
python-dotenv
azure-ai-formrecognizer
streamlit
